requirements.txt
# PyTorch Ecosystem (Install with CUDA support)
# Install separately first: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Core Scientific Computing
numpy>=2.1.0,<3.0.0
scipy>=1.11.0

# AI/ML Libraries - Latest Stable Versions
transformers>=4.57.0
diffusers>=0.35.0
accelerate>=1.12.0
safetensors>=0.4.0

# Computer Vision
opencv-python-headless>=4.8.0
imageio>=2.33.0
ultralytics>=8.3.233

# 3D Geometry & Visualization
meshlib>=3.0.0
viser>=0.2.0

# Configuration & Utilities
hydra-core>=1.3.2
loguru>=0.7.2